pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_code_reviewer"
version = "1.0.0"
description = "AI-Powered Code Reviewer & Quality Assistant"
authors = [{name = "AI Engineering Co-Pilot", email = "dev@example.com"}]
dependencies = [
    "streamlit>=1.24.0",
    "langchain-groq",
    "python-dotenv",
    "pydocstyle",
    "radon"
]
requires-python = ">=3.9"

[project.scripts]
ai-review = "ai_code_reviewer.cli:main"
ai-ui = "ai_code_reviewer.cli:run_ui"